+++ examples/resources/nxos_bfd/import-by-identity.tf
@@ -1,0 +1,4 @@
+import {
+  to       = nxos_bfd.example
+  identity = {}
+}